+++ src/main/java/io/github/wasiliystrecker/returns/SpringArchitecturePatternsApplication.java
@@ -5,9 +5,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableAsync
+@EnableScheduling
 public class SpringArchitecturePatternsApplication {
 
   public static void main(String[] args) {